Added car editing in user profile and separated activation letter sending
