+++ demoProj/config/services.yaml
@@ -11,6 +11,8 @@ services:
         public: false       # Allows optimizing the container by removing unused services; this also means
                             # fetching services directly from the container via $container->get() won't work.
                             # The best practice is to be explicit about your dependencies anyway.
+    app.email_activation_service:
+            class: App\Controller\EmailActivationController
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
